fix: Korrektheits- und Hygiene-Fixes für strenge Analyse-Gates
- bst.c: malloc-Null-Check + Baum am Ende freigeben (LeakSanitizer/cppcheck)
- verkettete_liste.c: malloc-Null-Check
- dijkstra.c: rev[] initialisieren (cppcheck uninitvar)
- eingabe-user-pruefen{,-v02}.c: %d -> %u für unsigned, getchar als int (EOF),
  Cast beim Speichern in char
- meineFkt.c: Vorzeichen-Cast in potenziere()
- halloC-v02.c: Konstanten double statt float
- integer.c: expliziter unsigned-Cast für UINT_MAX, abschließende Newline
- Trailing-Whitespace entfernt

Verhalten und Trace-Ausgabe bleiben unverändert.

+++ lib/meineFkt.c
@@ -34,11 +34,11 @@ int drittePotenz(int z){
 3 hoch 4 = 1* 3 * 3 * 3 * 3
          --- //trick
 3 hoch 0 = 1
-*/  
+*/
 unsigned long potenziere(int x, int y){ //basis, exponent
   unsigned long r = 1;
   for(int i = 0; i < y; i++){
-    r *= x;
+    r *= (unsigned long)x;   // x ist >= 0 (siehe Vertrag), Cast vermeidet Vorzeichen-Warnung
   }
   return r;
 }
